% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_organization.R
\name{ActiveIdent.gcdSeurat}
\alias{ActiveIdent.gcdSeurat}
\title{Finds and sets current active ident.}
\usage{
\method{ActiveIdent}{gcdSeurat}(object, ...)
}
\arguments{
\item{RET}{gcdSeurat object}
}
\value{
RET with meta.list modified to inclue active.ident
}
\description{
Finds and sets current active ident.
}
