% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_organization.R
\name{ActiveMarkers.gcdSeurat}
\alias{ActiveMarkers.gcdSeurat}
\title{Finds best markers available for current idents.}
\usage{
\method{ActiveMarkers}{gcdSeurat}(object, verbose = F, ...)
}
\arguments{
\item{RET}{gcdSeurat object}
}
\value{
markers slots containing entries matching ActiveIdent(RET)
}
\description{
Finds best markers available for current idents.
}
