% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCR_utils.R
\name{addTCRToSeurat}
\alias{addTCRToSeurat}
\title{Applies TCR annotation to Seurat object}
\usage{
addTCRToSeurat(object, barcode_to_clonotype, barcode_to_TRAV,
  barcode_to_TRBV)
}
\arguments{
\item{object}{Seurat object}

\item{barcode_to_clonotype}{list mapping barcode to clonotypes}

\item{barcode_to_TRAV}{list mapping barcodes to TRAV genes}

\item{barcode_to_TRBV}{list mapping barcodes to TRBV genes}
}
\value{
Seurat object with object$TCR_clonotype, object$TRA_V, 
and object$TRB_V filled in
}
\description{
Applies TCR annotation to Seurat object
}
\examples{
addTCRToSeurat(object, barcode_to_clonotype)

}
