% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCR_utils.R
\name{annotateCD3.CD20.cells}
\alias{annotateCD3.CD20.cells}
\title{Uses expression gating to label T cells.}
\usage{
annotateCD3.CD20.cells(object, CD20_cutoff = 0, CD3E_cutoff = 0)
}
\arguments{
\item{object}{Seurat object}

\item{CD20_cutoff}{threshold value for CD20+ (default is 0)}

\item{CD3E_cutoff}{threshold value for CD3+ (default is 0)}
}
\value{
Seurat object with object$CD4.CD8 filled in
}
\description{
Uses expression gating to label T cells.
}
\examples{
annotateCD4.CD8.TCells(object)

}
