% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCR_utils.R
\name{annotateCD4.CD8.TCells}
\alias{annotateCD4.CD8.TCells}
\title{Uses expression gating to label T cells.}
\usage{
annotateCD4.CD8.TCells(object, CD4_cutoff = 0.5, CD8A_cutoff = 0.5,
  CD3D_cutoff = 0.5)
}
\arguments{
\item{object}{Seurat object}

\item{CD4_cutoff}{threshold value for CD4+}

\item{CD8A_cutoff}{threshold value for CD8+}

\item{CD3D_cutoff}{threshold value for CD3+}
}
\value{
Seurat object with object$CD4.CD8 filled in
}
\description{
Uses expression gating to label T cells.
}
\examples{
annotateCD4.CD8.TCells(object)

}
