% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_organization.R
\name{BestMarkers.gcdSeurat}
\alias{BestMarkers.gcdSeurat}
\title{Finds best markers available for current idents.}
\usage{
\method{BestMarkers}{gcdSeurat}(object, ...)
}
\arguments{
\item{RET}{gcdSeurat object}
}
\value{
RET with meta.list modified to inclue active.ident
}
\description{
Finds best markers available for current idents.
}
