% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{blendedFeaturePlot}
\alias{blendedFeaturePlot}
\title{Applies the same X/Y axis limits to a list of plots.}
\usage{
blendedFeaturePlot(object, features, title = NULL, ...)
}
\arguments{
\item{...}{plots}
}
\description{
Applies the same X/Y axis limits to a list of plots.
}
