% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgsea_funcs.R
\name{checkFGSEARanks}
\alias{checkFGSEARanks}
\title{Wrapper running fgsea}
\usage{
checkFGSEARanks(RET, idents.to.use = NULL, overwrite = F)
}
\arguments{
\item{RET}{list containing Seurat object}

\item{idents.to.use}{}
}
\description{
Wrapper running fgsea
}
