% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_organization.R
\name{ClosestIdent}
\alias{ClosestIdent}
\title{Checks whether ident has been saved}
\usage{
\method{IdentsSaved}{gcdSeurat}(object, ...)
}
\arguments{
\item{object}{gcdSeurat}
}
\value{
whether idents must be saved
}
\description{
Checks whether ident has been saved
}
