% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomart_utils.R
\name{convertHumanGeneList}
\alias{convertHumanGeneList}
\title{Renames human genes to mouse}
\usage{
convertHumanGeneList(genes)
}
\arguments{
\item{genes}{list of human genes}
}
\value{
list containing mouse homologs of human genes
}
\description{
Renames human genes to mouse
}
\examples{
convertHumanGeneList(genes)

}
