% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomart_utils.R
\name{convertHumanPathwayList}
\alias{convertHumanPathwayList}
\title{Renames human genes to mouse}
\usage{
convertHumanPathwayList(pathways.list)
}
\arguments{
\item{pathways.list}{list of human genes}
}
\value{
list containing mouse homologs of human genes
}
\description{
Renames human genes to mouse
}
\examples{
convertHumanGeneList(genes)

}
