% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomart_utils.R
\name{convertMouseGeneList}
\alias{convertMouseGeneList}
\title{Renames mouse genes to human}
\usage{
convertMouseGeneList(de.table)
}
\arguments{
\item{genes}{list of mouse genes}
}
\value{
list containing human corresponding genes
}
\description{
Renames mouse genes to human
}
\examples{
convertMouseGeneList(genes)

}
