% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgsea_funcs.R
\name{correlateFgseaLeadingEdges}
\alias{correlateFgseaLeadingEdges}
\title{Correlates genes in leading edges of enriched genes}
\usage{
correlateFgseaLeadingEdges(RET, pval.thresh = 0.01,
  path.dbs.to.check = NULL)
}
\arguments{
\item{RET}{list containing Seurat object}

\item{pval.thresh}{significance threshold for plotting}

\item{path.dbs.to.check}{if given, limit plotting to only these pathway sets}
}
\value{
list containing ranks and fgsea output for each ident
}
\description{
Correlates genes in leading edges of enriched genes
}
\examples{
correlateFgseaLeadingEdges(RET, pval.thresh)

}
