% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_extensions.R
\name{correlateVariables}
\alias{correlateVariables}
\title{Correlates modules/genes}
\usage{
correlateVariables(RET, vars.to.corr, only.var = TRUE, split.by = NULL)
}
\arguments{
\item{RET}{list containing Seurat object}

\item{vars.to.corr}{variables to correlate ()}

\item{only.var}{if True, only consider named variables and not all modules}
}
\value{
list containing Seurat object and TSNE plots (tsne.treatment and tsne.ident) of the CCA
}
\description{
Correlates modules/genes
}
\examples{
correlateVariables(RET, vars.to.corr, True)

}
