% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCR_utils.R
\name{countGlycosylationSites}
\alias{countGlycosylationSites}
\title{Adds glycosylation site amino acid position and number to a clones data table.}
\usage{
countGlycosylationSites(clones.tbl)
}
\description{
Adds glycosylation site amino acid position and number to a clones data table.
}
