% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgsea_funcs.R
\name{fgseaFromMarkers}
\alias{fgseaFromMarkers}
\title{Wrapper running fgsea}
\usage{
fgseaFromMarkers(markers, pathways.list, prefixes = NULL)
}
\arguments{
\item{markers}{DE marker table}

\item{pathways.list}{list of pathway sets to analyze}

\item{prefixes}{list of prefixes to trim from name of each pathway}
}
\value{
list containing ranks and fgsea output for each ident
}
\description{
Wrapper running fgsea
}
\examples{
fgseaFromMarkers(markers, pathways.list)

}
