% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgsea_funcs.R
\name{fgseaWrapper}
\alias{fgseaWrapper}
\title{Wrapper running fgsea}
\usage{
fgseaWrapper(RET, pathways.list, prefixes = NULL, mouse = FALSE)
}
\arguments{
\item{RET}{list containing Seurat object}

\item{pathways.list}{list of pathway sets to analyze}

\item{prefixes}{list of prefixes to trim from name of each pathway}

\item{mouse}{whether to convert gene names prior to fgsea}
}
\value{
list containing ranks and fgsea output for each ident
}
\description{
Wrapper running fgsea
}
\examples{
fgseaWrapper(RET, pathways.list)

}
