% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_scoring.R
\name{filterModules}
\alias{filterModules}
\title{Filters modules}
\usage{
filterModules(modules, control.pool)
}
\arguments{
\item{modules}{named list of modules (containing gene names) to be added}
}
\description{
Filters modules
}
