% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgsea_funcs.R
\name{findEnrichedPathways}
\alias{findEnrichedPathways}
\title{Finds enriched pathways using fgsea}
\usage{
findEnrichedPathways(RET, dbs)
}
\arguments{
\item{RET}{list containing Seurat object}

\item{dbs}{pathway databases}
}
\value{
list containing Seurat object and marker heatmaps
}
\description{
Finds enriched pathways using fgsea
}
\examples{
makeMarkerHeatmaps(RET, marker.lists)

}
