% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_wrappers.R
\name{gcdPrintSeuratQC}
\alias{gcdPrintSeuratQC}
\title{Prints filtering info for a gcdSeurat object, describing how many cells pass each filter.}
\usage{
gcdPrintSeuratQC(RET)
}
\arguments{
\item{RET}{Seurat object}
}
\description{
Prints filtering info for a gcdSeurat object, describing how many cells pass each filter.
}
\examples{
seuratFilterWrapper(SRT, min.genes = min.genes, max.genes = max.genes, max.UMI = max.UMI,
   max.mito = max.mito, mito.prefix = mito.prefix)

}
