% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_input.R
\name{gcdRead10X}
\alias{gcdRead10X}
\title{Reads in 10x feature data}
\usage{
gcdRead10X(data.dir = NULL, gene.column = 2)
}
\arguments{
\item{data.dir}{same as regular 10x arg}

\item{gene.column}{same as regular 10x arg}
}
\value{
none
}
\description{
Reads in 10x feature data
}
\examples{
counts <- gcdRead10X("Lib_A_GRC_GRCh38v3.0.1")
print(counts$`Gene Expression`)

}
