% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_organization.R
\name{gcdRenameIdent}
\alias{gcdRenameIdent}
\title{Renames selected groups of cells in a Seurat object}
\usage{
gcdRenameIdent(RET, ident.to.rename, new.label)
}
\arguments{
\item{RET}{list containing Seurat object and plots}

\item{ident.to.rename}{SINGLE ident to rename}

\item{new.label}{new name for ident}
}
\value{
list containing renamed Seurat object and re-plotted TSNE
}
\description{
Renames selected groups of cells in a Seurat object
}
\examples{
renameIdents(RET)

}
