% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_wrappers.R
\name{gcdRenameIdents}
\alias{gcdRenameIdents}
\title{Renames selected groups of cells in a Seurat object}
\usage{
gcdRenameIdents(RET, idents.to.rename, new.label)
}
\arguments{
\item{RET}{list containing Seurat object and plots}

\item{idents.to.rename}{list containing idents to rename}

\item{new.idents}{new name for idents}
}
\value{
list containing renamed Seurat object and re-plotted TSNE
}
\description{
Renames selected groups of cells in a Seurat object
}
\examples{
renameIdents(RET)

}
