% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_organization.R
\docType{class}
\name{gcdSeurat}
\alias{gcdSeurat}
\title{gcdSeurat}
\description{
Replacing the simple list I had before with this, hopefully improving memory usage.
}
\section{Slots}{

\describe{
\item{\code{seurat}}{Unnormalized data such as raw counts or TPMs}

\item{\code{meta.list}}{Normalized expression data}
}}

