% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_extensions.R
\name{gcdSeuratCellCycleScoring}
\alias{gcdSeuratCellCycleScoring}
\title{Performs cell cycle scoring}
\usage{
gcdSeuratCellCycleScoring(object, s.features, g2m.features, set.ident = FALSE)
}
\arguments{
\item{object}{Seurat object}

\item{s.features}{list of features (genes) associated with S phase}

\item{g2m.features}{list of features (genes) associated with G2/M phase}

\item{set.ident}{whether to set idents of Seurat object}
}
\value{
list containing Seurat object and TSNE plots (tsne.treatment and tsne.ident) of the CCA
}
\description{
Performs cell cycle scoring
}
\examples{
gcdCellCycleScoring(RET, modules)

}
