% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoupX_load.R
\name{gcdSoupX.load10X}
\alias{gcdSoupX.load10X}
\title{Reads in 10x feature data}
\usage{
gcdSoupX.load10X(
  dataDir,
  cellIDs = NULL,
  channelName = NULL,
  readArgs = list(),
  includeFeatures = c("Gene Expression"),
  verbose = TRUE,
  require.intersect = T...
)
}
\arguments{
\item{dataDir}{Top level cellranger output directory (the directory that contains the \code{raw_gene_bc_matrices} folder).}

\item{cellIDs}{Barcodes of droplets that contain cells.  If NULL, use the default cellranger set.}

\item{channelName}{The name of the channel to store.  If NULL set to either \code{names(dataDir)} or \code{dataDir} is no name is set.}

\item{readArgs}{A list of extra parameters passed to \code{Seurat::Read10X}.}

\item{includeFeatures}{If multiple feature types are present, keep only the types mentioned here and collapse to a single matrix.}

\item{verbose}{Be verbose?}

\item{require.intersect}{If TRUE, stop if cellIDs do not overlap perfectly between metadata and filtered data. If FALSE, warn if cells are missing from either, but proceed with the intersection of cells.}

\item{data.dir}{same as regular 10x arg}

\item{gene.column}{same as regular 10x arg}

\item{trim.suffix}{Remove the "-1" suffix if present in cell barcodes?}

\item{...}{Extra parameters passed to \code{SoupChannel} construction function.}
}
\value{
none

A SoupChannel object containing the count tables for the 10X dataset.
}
\description{
Reads in 10x feature data
}
\examples{
counts <- gcdRead10X("Lib_A_GRC_GRCh38v3.0.1")
print(counts$`Gene Expression`)

sc = load10X(system.file('extdata','toyData',package='SoupX'))
}
\seealso{
SoupChannel estimateSoup
}
