% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_wrappers.R
\name{gcdSubsetSeurat}
\alias{gcdSubsetSeurat}
\title{Subsets a gcdSeurat object}
\usage{
gcdSubsetSeurat(RET, ...)
}
\arguments{
\item{RET}{list containing Seurat object and plots}

\item{cells}{cells to subset}
}
\value{
subsetted gcdSeurat
}
\description{
Subsets a gcdSeurat object
}
\examples{
gcdSubsetSeurat(RET)

}
