% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_organization.R
\name{gcdUpdateSeurat}
\alias{gcdUpdateSeurat}
\title{Reads Seurat/gcdSeurat object and converts it to updated gcdSeurat}
\usage{
gcdUpdateSeurat(RET = NULL, SRT = NULL, shiny = TRUE)
}
\arguments{
\item{RET}{gcdSeurat object}
}
\value{
RET with meta.list modified to inclue active.ident
}
\description{
Reads Seurat/gcdSeurat object and converts it to updated gcdSeurat
}
