% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_extensions.R
\name{getCorrLists}
\alias{getCorrLists}
\title{Helper function for correlateVariables}
\usage{
getCorrLists(vars.to.corr, exp_matrix, ident.name = "All cells")
}
\description{
Helper function for correlateVariables
}
