% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgsea_funcs.R
\name{gseaPerCluster}
\alias{gseaPerCluster}
\title{Finds enriched pathways in each cluster using fgsea}
\usage{
gseaPerCluster(all.markers, pathways)
}
\arguments{
\item{all.markers}{table of DE genes between conditions, returned by \code{\link[Seurat]{FindAllMarkers}}}

\item{dbs}{pathway databases}
}
\value{
list containing ranks and fgsea output for each ident
}
\description{
Finds enriched pathways in each cluster using fgsea
}
\examples{
gseaPerCluster(all.markers, pathways)

}
