% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_organization.R
\name{GseaRes.gcdSeurat}
\alias{GseaRes.gcdSeurat}
\title{Returns fgsea results for current idents.}
\usage{
\method{GseaRes}{gcdSeurat}(object, ...)
}
\arguments{
\item{RET}{gcdSeurat object}
}
\value{
RET with meta.list modified to inclue active.ident
}
\description{
Returns fgsea results for current idents.
}
