% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgsea_funcs.R
\name{gseaTopPlots}
\alias{gseaTopPlots}
\title{Plots enriched and depleted pathways in each cluster}
\usage{
gseaTopPlots(
  fgseaRes,
  path.name,
  pathways,
  term.prefix = "GO_",
  pval.thresh = 0.01,
  num.to.print = 10
)
}
\arguments{
\item{fgseaRes}{list containing ranks and fgsea output for each ident, output of \link{gseaPerCluster}}

\item{path.name}{name of pathway database}

\item{pathways}{list of pathways}

\item{term.prefix}{prefix to remove from pathway name for printing}

\item{pval.thresh}{only include pathways with padj below this value}

\item{num.to.print}{print at most this many pathways}
}
\value{
list containing ranks and fgsea output for each ident
}
\description{
Plots enriched and depleted pathways in each cluster
}
\examples{
gseaTopPlots(all.markers, pathways)

}
