% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_extensions.R
\name{guessGCDSeuratSpecies}
\alias{guessGCDSeuratSpecies}
\title{Guesses whether data is aligned to a human, mouse, or other reference.
Uses msigdbr gene lists.}
\usage{
guessGCDSeuratSpecies(RET, exhaustive = F, verbose = F)
}
\arguments{
\item{RET}{gcdSeurat}

\item{exhaustive}{FALSE by default, meaning only mouse and human are considered. Use TRUE to check all possible species in msigdbr.}
}
\value{
best fit species
}
\description{
Guesses whether data is aligned to a human, mouse, or other reference.
Uses msigdbr gene lists.
}
\examples{
guessGCDSeuratSpecies(RET)

}
