% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_extensions.R
\name{guessSpeciesAddCellCycleScoring}
\alias{guessSpeciesAddCellCycleScoring}
\title{Adds cell cycle scoring to mouse/human scRNAseq data}
\usage{
guessSpeciesAddCellCycleScoring(RET, overwrite = F, shiny = T)
}
\arguments{
\item{RET}{gcdSeurat object}
}
\value{
RET with meta.list modified to inclue active.ident
}
\description{
Adds cell cycle scoring to mouse/human scRNAseq data
}
