% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_organization.R
\name{CellGroups.gcdSeurat}
\alias{CellGroups.gcdSeurat}
\title{Renames selected groups of cells in a Seurat object}
\usage{
\method{CellGroups}{gcdSeurat}(object, ...)
}
\description{
Renames selected groups of cells in a Seurat object
}
