% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCR_utils.R
\name{igphyml.treePlot}
\alias{igphyml.treePlot}
\title{Dendrogram plot}
\usage{
igphyml.treePlot(tree, title.text = "", cells.highlight = NULL,
  cols.highlight = NULL, highlight.name = "Highlight",
  highlight.legend = T, highlight.legend.pos = "topright",
  site.colors = NULL, ymin = NULL, ymax = NULL, highlight.leaf = F,
  leaf.pch = 17, leaf.cex = 2, show.pbmc = T, show.timepoint = F,
  hide.yaxt = T)
}
\description{
Dendrogram plot
}
