% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_exp.R, R/seurat_extensions.R
\name{markersBetweenConditions}
\alias{markersBetweenConditions}
\title{Finds DE genes in each cluster between conditions}
\usage{
markersBetweenConditions(RET, cond.var, cond.1, cond.2)

markersBetweenConditions(RET, cond.var, cond.1, cond.2)
}
\arguments{
\item{RET}{list containing aligned Seurat object}

\item{cond.var}{name of variable storing condition values in Seurat object's metadata.}

\item{cond.1}{first condition}

\item{cond.2}{second condition}
}
\value{
dataframe containing DE markers between conditions

dataframe containing DE markers between conditions
}
\description{
Finds DE genes in each cluster between conditions

Finds DE genes in each cluster between conditions
}
\examples{
markersBetweenConditions(RET)

markersBetweenConditions(RET)

}
