% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mouse.cc.genes}
\alias{mouse.cc.genes}
\title{Mouse cell cycle genes}
\format{An object of class \code{list} of length 2.}
\usage{
data(mouse.cc.genes)
}
\description{
Mouse cell cycle genes
}
\examples{
data(mouse.cc.genes)
}
\references{
Liberzon et al. The Molecular Signatures Database (MSigDB) hallmark gene set collection. Cell Syst. 2015 Dec 23;1(6):417-425. 
(\href{https://www.ncbi.nlm.nih.gov/pubmed/26771021}{PubMed})
}
\keyword{datasets}
