% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSIGDB.pathways.R
\docType{data}
\name{MSIGDB.pathways}
\alias{MSIGDB.pathways}
\title{MSIGDB pathway lists}
\format{An object of class \code{list} of length 7.}
\usage{
data(MSIGDB.pathways)
}
\description{
MSIGDB pathway lists
}
\examples{
data(MSIGDB.pathways)
}
\references{
Liberzon et al. The Molecular Signatures Database (MSigDB) hallmark gene set collection. Cell Syst. 2015 Dec 23;1(6):417-425. 
(\href{https://www.ncbi.nlm.nih.gov/pubmed/26771021}{PubMed})
}
\keyword{datasets}
