% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCR_utils.R
\name{parseBuildTreesLogfile}
\alias{parseBuildTreesLogfile}
\title{Adds glycosylation site amino acid position and number to a clones data table.}
\usage{
parseBuildTreesLogfile(file.path, patient = NULL)
}
\description{
Adds glycosylation site amino acid position and number to a clones data table.
}
