% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_wrappers.R
\name{percent.table}
\alias{percent.table}
\title{Prints table as percentage values}
\usage{
percent.table(tbl)
}
\arguments{
\item{tbl}{table}
}
\value{
none
}
\description{
Prints table as percentage values
}
\examples{
RET <- makeMarkerHeatmaps(RET, marker.lists)
printMarkerHeatmaps(RET)

}
