% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCR_utils.R
\name{plotClonotypeCD4.CD8}
\alias{plotClonotypeCD4.CD8}
\title{Dimensional reduction plot highlighting cells of a given clonotype based on their CD4.CD8 status}
\usage{
plotClonotypeCD4.CD8(object, title, clonotype, clonotype_prefix, ...)
}
\description{
Dimensional reduction plot highlighting cells of a given clonotype based on their CD4.CD8 status
}
