% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCR_utils.R
\name{plotClonotypeIntVsSCT}
\alias{plotClonotypeIntVsSCT}
\title{Pair of dimensional reduction plots highlighting cells of a given clonotype}
\usage{
plotClonotypeIntVsSCT(
  object1,
  prefix1,
  object2,
  prefix2,
  clonotype,
  clonotype_prefix,
  clonotype_field = "TCR_clonotype"
)
}
\description{
Pair of dimensional reduction plots highlighting cells of a given clonotype
}
