% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCR_utils.R
\name{plotClonotypeTimePoints}
\alias{plotClonotypeTimePoints}
\title{Dimensional reduction plot highlighting cells of a given clonotype based on their CD4.CD8 status}
\usage{
plotClonotypeTimePoints(object, title, clonotype, clonotype_prefix,
  timepoint.field = "timepoint", site.field = "site",
  site.levels = NULL, timepoint.levels = NULL, ...)
}
\description{
Dimensional reduction plot highlighting cells of a given clonotype based on their CD4.CD8 status
}
