% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotWithSameXYLimits}
\alias{plotWithSameXYLimits}
\title{Applies the same X/Y axis limits to a list of plots.}
\usage{
plotWithSameXYLimits(..., plotlist = NULL)
}
\arguments{
\item{...}{plots}
}
\description{
Applies the same X/Y axis limits to a list of plots.
}
