% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_wrappers.R
\name{prettyPrintMarkers}
\alias{prettyPrintMarkers}
\title{Creates table of marker genes differentially expressed within cells belonging to ident}
\usage{
prettyPrintMarkers(markers, ident)
}
\arguments{
\item{markers}{data frame returned by \code{\link[Seurat]{FindAllMarkers}}}

\item{ident}{cluster/group of interest}
}
\value{
Well-formatted data frame
}
\description{
Creates table of marker genes differentially expressed within cells belonging to ident
}
\examples{
prettyPrintMarkers()

}
