% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgsea_funcs.R
\name{printGseaPlots}
\alias{printGseaPlots}
\title{Prints fgsea plots for Seurat object}
\usage{
printGseaPlots(RET, redraw = F, path.dbs.to.check = NULL, ...)
}
\arguments{
\item{RET}{list containing Seurat object}

\item{redraw}{whether plots should be redrawn before plotting}

\item{path.dbs.to.check}{if given, limit plotting to only these pathway sets}
}
\value{
list containing ranks and fgsea output for each ident
}
\description{
Prints fgsea plots for Seurat object
}
\examples{
printGseaPlots(RET, pathways.list)

}
