% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_wrappers.R
\name{printMarkerHeatmaps}
\alias{printMarkerHeatmaps}
\title{Prints heatmaps for Seurat object}
\usage{
printMarkerHeatmaps(RET)
}
\arguments{
\item{RET}{list containing Seurat object and plots}
}
\value{
none
}
\description{
Prints heatmaps for Seurat object
}
\examples{
RET <- makeMarkerHeatmaps(RET, marker.lists)
printMarkerHeatmaps(RET)

}
