% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{printSeuratMarkers}
\alias{printSeuratMarkers}
\title{Prints markers distinguishing clusters in a Seurat object}
\usage{
printSeuratMarkers(RET)
}
\arguments{
\item{RET}{list containing Seurat object and plots}
}
\value{
list containing clustered Seurat object and TSNE plots
}
\description{
Prints markers distinguishing clusters in a Seurat object
}
\examples{
printSeuratMarkers(RET)

}
