% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_utilities.R
\name{readMarkersYAML}
\alias{readMarkersYAML}
\title{Reads in a list of markers by cell type}
\usage{
readMarkersYAML(yaml.path)
}
\arguments{
\item{RET}{list containing Seurat object and plots}
}
\value{
none
}
\description{
Reads in a list of markers by cell type
}
\examples{
readMarkersYAML(yaml.path)


}
