% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_organization.R
\name{readSeuratRDStoGCD}
\alias{readSeuratRDStoGCD}
\title{Reads Seurat/gcdSeurat object and converts it to updated gcdSeurat}
\usage{
readSeuratRDStoGCD(path = "", object = NULL, shiny = FALSE)
}
\arguments{
\item{path}{Filepath to .rds object storing a gcdSeurat or Seurat object}

\item{object}{In-memory gcdSeurat/Seurat object.}

\item{shiny}{Print shiny progress updates?}
}
\value{
gcdSeurat object
}
\description{
Reads Seurat/gcdSeurat object and converts it to updated gcdSeurat
}
