% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_wrappers.R
\name{renameIdents}
\alias{renameIdents}
\title{Renames idents of cells in a Seurat object}
\usage{
renameIdents(RET, new.idents)
}
\arguments{
\item{RET}{list containing Seurat object and plots}

\item{new.idents}{list containing new idents (in same order as current)}
}
\value{
list containing renamed Seurat object and re-plotted TSNE
}
\description{
Renames idents of cells in a Seurat object
}
\examples{
renameIdents(RET)

}
