% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_viewer.R
\name{runCellViewer}
\alias{runCellViewer}
\title{Launches Seurat viewer app}
\usage{
runCellViewer(object = NULL, ...)
}
\arguments{
\item{...}{Arguments passed to shiny::runApp}
}
\description{
Launches Seurat viewer app
}
\examples{
runCellViewer()

}
