% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgsea_funcs.R
\name{runEnrichrOnDEGenes}
\alias{runEnrichrOnDEGenes}
\title{Runs DE genes through enrichR}
\usage{
runEnrichrOnDEGenes(RET, path.dbs.to.check,
  pval.thresh = 1.00000000000000e-50, logfc.thresh = 0,
  idents.to.use = NULL)
}
\arguments{
\item{RET}{gcdSeurat}

\item{path.dbs.to.check}{if given, limit plotting to only these pathway sets}

\item{pval.thresh}{threshold p-val for inclusion in DE gene list}

\item{logfc.thresh}{threshold logfc-val for inclusion in DE gene list}
}
\value{
list containing ranks and fgsea output for each ident
}
\description{
Runs DE genes through enrichR
}
\examples{
correlateFgseaLeadingEdges(RET, pval.thresh)

}
