% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_scoring.R
\name{scoreModules}
\alias{scoreModules}
\title{Adds module scores to a Seurat object}
\usage{
scoreModules(RET, modules, verbose = T, min.cells = 5)
}
\arguments{
\item{RET}{list containing Seurat object}

\item{modules}{named list of modules (containing gene names) to be added}
}
\value{
list containing Seurat object and TSNE plots (tsne.treatment and tsne.ident) of the CCA
}
\description{
Adds module scores to a Seurat object
}
\examples{
scoreModules(RET, modules)

}
