% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCR_utils.R
\name{separateCellnamesVDJ}
\alias{separateCellnamesVDJ}
\title{Splits cell name barcodes into usable dataframe.}
\usage{
separateCellnamesVDJ(tree = NULL, cellnames = NULL)
}
\value{
dataframe with relevant fields: notably $fixed_cellnames,
containing properly formatted cell barcodes.
}
\description{
Splits cell name barcodes into usable dataframe.
}
