% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_organization.R
\name{setActiveIdent}
\alias{setActiveIdent}
\title{Finds and sets current active ident.}
\usage{
setActiveIdent(RET)
}
\arguments{
\item{RET}{gcdSeurat object}
}
\value{
RET with meta.list modified to inclue active.ident
}
\description{
Finds and sets current active ident.
}
