% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_exp.R
\name{setMarkersByCluster}
\alias{setMarkersByCluster}
\title{Adds markers to cluster list}
\usage{
setMarkersByCluster(RET, markers.type, markers)
}
\arguments{
\item{RET}{gcdSeurat}

\item{markers.type}{name for markers (e.g. "quick" or "full")}

\item{markers}{returned data frame from Seurat FindAllMarkers}
}
\value{
gcdSeurat
}
\description{
Adds markers to cluster list
}
\examples{
setMarkersByCluster(RET, markers.type, markers)

}
