% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_CCA_wrappers.R
\name{seuratAlignClusterWrapper}
\alias{seuratAlignClusterWrapper}
\title{Aligns clusters between two conditions}
\usage{
seuratAlignClusterWrapper(RET, dims.align = 1:20, resolution = 0.6)
}
\arguments{
\item{RET}{list containing Seurat objects for cond1 and cond2 and plots}

\item{dims.align}{dimensions to be used for aligment and clustering}

\item{resolution}{resolution to be used for clustering}
}
\value{
list containing aligned Seurat object and TSNE plot of the CCA
}
\description{
Aligns clusters between two conditions
}
\examples{
seuratAlignClusterWrapper(RET)

}
