% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_wrappers.R, R/diff_exp.R
\name{seuratAllMarkers}
\alias{seuratAllMarkers}
\title{Wrapper function to find markers between clusters for a Seurat object}
\usage{
seuratAllMarkers(RET, do.fast = TRUE, do.full = FALSE)

seuratAllMarkers(RET, do.fast = TRUE, do.full = FALSE)
}
\arguments{
\item{RET}{list containing Seurat object and plots}

\item{do.fast}{whether fast markers should be computed (default: True)}

\item{do.full}{whether full markers should be computed (default: True)}

\item{RET}{list containing Seurat object and plots}

\item{do.fast}{whether fast markers should be computed (default: True)}

\item{do.full}{whether full markers should be computed (default: True)}
}
\value{
list containing clustered Seurat object and TSNE plots

list containing clustered Seurat object and TSNE plots
}
\description{
Wrapper function to find markers between clusters for a Seurat object

Wrapper function to find markers between clusters for a Seurat object
}
\examples{
seuratAllMarkers(RET)

seuratAllMarkers(RET)

}
