% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgsea_funcs.R
\name{seuratDEresToGSEARanks}
\alias{seuratDEresToGSEARanks}
\title{Convers Seurat marker output to DE gene ranks}
\usage{
seuratDEresToGSEARanks(de.table, p_thresh = 1e-50, logfc_thresh = 0.25)
}
\arguments{
\item{de.table}{DE genes table, returned by \code{\link[Seurat]{FindAllMarkers}}}
}
\value{
ranks to be used for \code{\link{gseaPerCluster}}
}
\description{
Convers Seurat marker output to DE gene ranks
}
\examples{
seuratDEresToGSEARanks(de.table)

}
