% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_extensions.R
\name{seuratFindLikelyLightChainIdent}
\alias{seuratFindLikelyLightChainIdent}
\title{Identifies light chain identity of sample}
\usage{
seuratFindLikelyLightChainIdent(SRT, mouse = FALSE,
  tumor.idents = NULL)
}
\arguments{
\item{SRT}{Seurat object}

\item{mouse}{whether mouse (e.g. Igh...) or human (IGH...) gene names should be assumed}

\item{tumor.idents}{identities of clusters containing tumor cells}
}
\value{
IG.light.V, IG.light.C
}
\description{
Identifies light chain identity of sample
}
\examples{
RET <- findLikelyLightChainIdent(RET)

}
