% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_wrappers.R
\name{seuratMarkersBetweenConditions}
\alias{seuratMarkersBetweenConditions}
\title{Finds conserved markers for all clusters}
\usage{
seuratMarkersBetweenConditions(RET)
}
\arguments{
\item{RET}{list containing aligned Seurat object}
}
\value{
list containing aligned Seurat object and TSNE plot of the CCA
}
\description{
Finds conserved markers for all clusters
}
\examples{
seuratAlignClusterWrapper(RET)

}
