% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_exp.R
\name{seuratModuleMarkers}
\alias{seuratModuleMarkers}
\title{Wrapper function to find module markers between clusters for a Seurat object}
\usage{
seuratModuleMarkers(RET, modules.use = NULL, markers.name = "modules",
  idents.use = NULL, shiny = F, logfc.threshold = 0.05)
}
\arguments{
\item{RET}{list containing Seurat object and plots}

\item{do.fast}{whether fast markers should be computed (default: True)}

\item{do.full}{whether full markers should be computed (default: True)}

\item{fast.threshold}{whether full markers should be computed (default: True)}
}
\value{
gcdSeurat
}
\description{
Wrapper function to find module markers between clusters for a Seurat object
}
\examples{
seuratModuleMarkers(RET)

}
