% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_wrappers.R
\name{seuratPCAWrapper}
\alias{seuratPCAWrapper}
\title{Wrapper function to perform PCA and generate PCA plots for a Seurat object}
\usage{
seuratPCAWrapper(RET, do.jackstraw = FALSE)
}
\arguments{
\item{RET}{list containing Seurat object and plots}

\item{do.jackstraw}{whether JackStraw analysis is performed}
}
\value{
list containing post-PCA Seurat object and PCA plots
}
\description{
Wrapper function to perform PCA and generate PCA plots for a Seurat object
}
\examples{
seuratPCAWrapper(RET)

}
