% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_CCA_wrappers.R
\name{seuratTSNEPlotCCA}
\alias{seuratTSNEPlotCCA}
\title{Makes TSNE plot after CCA}
\usage{
seuratTSNEPlotCCA(RET)
}
\arguments{
\item{RET}{list containing aligned Seurat object}
}
\value{
list containing Seurat object and TSNE plots (tsne.treatment and tsne.ident) of the CCA
}
\description{
Makes TSNE plot after CCA
}
\examples{
seuratTSNEPlotCCA(RET)

}
