% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_wrappers.R
\name{seuratVariableWrapper}
\alias{seuratVariableWrapper}
\title{Wrapper function to find variable genes in Seurat object}
\usage{
seuratVariableWrapper(RET, nfeatures = 2500, vars.to.regress = NULL,
  do.normalize = F, scale.all.features = TRUE)
}
\arguments{
\item{RET}{list containing Seurat object and plots}
}
\value{
list containing scaled Seurat object and plot of variable genes
}
\description{
Wrapper function to find variable genes in Seurat object
}
\examples{
seuratVariableWrapper(RET)

}
