% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{splitBarPlot}
\alias{splitBarPlot}
\title{Applies the same X/Y axis limits to a list of plots.}
\usage{
splitBarPlot(object, group.by, split.by, cols, do.percent = T, do.facet = F)
}
\arguments{
\item{...}{plots}
}
\description{
Applies the same X/Y axis limits to a list of plots.
}
