% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_utilities.R
\name{SummarizeExprAcrossIdentsOld}
\alias{SummarizeExprAcrossIdentsOld}
\title{Calculates average expression and counts cells with 
nonzero expression for each ident.}
\usage{
SummarizeExprAcrossIdentsOld(object, assay = NULL, features = NULL,
  slot = "data", verbose = F)
}
\arguments{
\item{object}{A Seurat object}

\item{assay}{Assay to use in differential expression testing}

\item{slot}{Slot to pull data from}

\item{verbose}{Print message when starting calculation for each ident}
}
\value{
List of two matrices: log_expr, containing mean of exmp1-transformed 
expression values for each feature across the cells in each ident; 
and expr_counts, containing the count of cells in each ident expressing 
each feature
}
\description{
Calculates average expression and counts cells with 
nonzero expression for each ident.
}
\examples{
\dontrun{
ident.summary <- SummarizeExprAcrossIdents(pbmc_small)
}

}
