% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_organization.R
\name{tryOverwriteIdent}
\alias{tryOverwriteIdent}
\title{Finds and sets current active ident.}
\usage{
tryOverwriteIdent(RET, OVERWRITE.IDENT)
}
\arguments{
\item{RET}{gcdSeurat object}
}
\value{
RET with meta.list modified to inclue active.ident
}
\description{
Finds and sets current active ident.
}
