% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_organization.R
\name{updateSeuratAndRunUMAP}
\alias{updateSeuratAndRunUMAP}
\title{Reads Seurat/gcdSeurat object and converts it to updated gcdSeurat}
\usage{
updateSeuratAndRunUMAP(RET, shiny = T)
}
\arguments{
\item{RET}{gcdSeurat object}
}
\value{
RET with meta.list modified to inclue active.ident
}
\description{
Reads Seurat/gcdSeurat object and converts it to updated gcdSeurat
}
