% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{breakdownTable}
\alias{breakdownTable}
\title{Summarizes 2 variables}
\usage{
breakdownTable(var.1, var.2, transpose = F, do.percent = F)
}
\arguments{
\item{var.1}{first list of values with length n}

\item{var.2}{second list of values with length m}

\item{transpose}{whether to flip results}

\item{do.percent}{whether to use percent or raw counts (default)}
}
\description{
Summarizes 2 variables
}
