% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgsea_funcs.R
\name{findLikelyLightChainIdent}
\alias{findLikelyLightChainIdent}
\title{Identifies light chain identity of sample}
\usage{
findLikelyLightChainIdent(RET, sample.variable = NULL)
}
\arguments{
\item{RET}{list containing Seurat object}

\item{sample.variable}{name distinguishing sample idents}
}
\value{
updated \code{RET}
}
\description{
Identifies light chain identity of sample
}
\examples{
RET <- findLikelyLightChainIdent(RET)

}
