% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCR_utils.R
\name{parseBuildTreesLogfile}
\alias{parseBuildTreesLogfile}
\title{Parses the output of BuildTrees (ChangeO function)
Returns a dataframe assigning each record to a subclone}
\usage{
parseBuildTreesLogfile(file.path, patient = NULL, debug = F)
}
\description{
Parses the output of BuildTrees (ChangeO function)
Returns a dataframe assigning each record to a subclone
}
