% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{readMarkersYAML}
\alias{readMarkersYAML}
\title{Prints heatmaps for Seurat object}
\usage{
readMarkersYAML(yaml.path)
}
\arguments{
\item{RET}{list containing Seurat object and plots}
}
\value{
none
}
\description{
Prints heatmaps for Seurat object
}
\examples{
RET <- makeMarkerHeatmaps(RET, marker.lists)
printMarkerHeatmaps(RET)

}
