% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_exp.R
\name{seuratAllMarkers}
\alias{seuratAllMarkers}
\title{Wrapper function to find markers between clusters for a Seurat object}
\usage{
seuratAllMarkers(RET, do.fast = TRUE, do.full = FALSE,
  fast.threshold = 0.25, shiny = F, idents.use = NULL,
  assay = NULL, slot = "data", features = NULL, verbose = T,
  test.use = "wilcox", ident.summary.data = NULL, ...)
}
\arguments{
\item{RET}{list containing Seurat object and plots}

\item{do.fast}{whether fast markers should be computed (default: True)}

\item{do.full}{whether full markers should be computed (default: True)}

\item{fast.threshold}{whether full markers should be computed (default: True)}
}
\value{
list containing clustered Seurat object and TSNE plots
}
\description{
Wrapper function to find markers between clusters for a Seurat object
}
\examples{
seuratAllMarkers(RET)

}
