% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgsea_funcs.R
\name{seuratFindLikelyLightChainIdent}
\alias{seuratFindLikelyLightChainIdent}
\title{Identifies light chain identity of sample}
\usage{
seuratFindLikelyLightChainIdent(SRT)
}
\arguments{
\item{SRT}{Seurat object}
}
\value{
IG.light.V, IG.light.C
}
\description{
Identifies light chain identity of sample
}
\examples{
RET <- findLikelyLightChainIdent(RET)

}
