% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_extensions.R
\name{seuratMarkersBetweenConditions}
\alias{seuratMarkersBetweenConditions}
\title{Finds DE genes in each cluster between conditions}
\usage{
seuratMarkersBetweenConditions(SRT, cond.var, cond.1, cond.2)
}
\arguments{
\item{SRT}{list containing aligned Seurat object}

\item{cond.var}{name of variable storing condition values in Seurat object's metadata.}

\item{cond.1}{first condition}

\item{cond.2}{second condition}
}
\value{
dataframe containing DE markers between conditions
}
\description{
Finds DE genes in each cluster between conditions
}
\examples{
seuratMarkersBetweenConditions(SRT)

}
